Add default names to the open_mat function
The default names makes it easier to use the function because the user
does not have to add ['Travel', 'Travelrate', 'Pitch', 'Pitchrate',
'Elevation', 'Elevationrate'] to every call of open_mat.

To allow longer lists there is an input "names_ext" that is concatenated
with the names list.

It is also possible to override the names if desired by simply adding
names to the arguments.