Add workaround to enable plotting in ncols!=1
This requires the user to define their own fig and ax!

You can now pass fig and ax to plot comparisons, and force xlabels to
get "time[s]" on the bottom of each column.

This is very much a workaround, but as this only needs to work for LAB4
it's fine <3